requirements.txt
# Course-to-Agents pipeline: Gemini (video → JSON) + Claude (JSON → code) + LangGraph
google-genai>=1.0.0
anthropic>=0.39.0
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-anthropic>=0.2.0
python-dotenv>=1.0.0
# Tool bindings (Instantly, HeyGen, Stripe)
requests>=2.28.0
stripe>=10.0.0
# Deployment (24/7 server)
uvicorn[standard]>=0.30.0
fastapi>=0.115.0

# Browser UI (Lovable-style step-by-step)
chainlit>=1.0.0

# Persistence for Chainlit (survives reload)
langgraph-checkpoint-sqlite>=2.0.0
